feat(otel): add redis topics and mechanisms and OTEL diagrams explaining new service's specifications (#1882)
* feat(sinks/sinker): wip add redis topic to handle config yaml.


* feat(otel): add diagrams to explain new head on otel.


* feat(otel): add create diagram.


* feat(otel): fix missing dataset.


* feat(otel): update diagrams


* feat(sinks/sinker): wip add redis topic to handle config yaml.


* feat(maestro): wip, need to fetch config from redis cache


* feat(maestro): wip adjusting to each event and fixing collector creation/update on Sinks, only in sinker update.


* feat(maestro): wip adjusting to each event and fixing collector creation/update on Sinks, only in sinker update.


* feat(maestro): wip adjusting to each event and fixing collector creation/update on Sinks, only in sinker update.


* feat(maestro): wip adjusting to each event and fixing collector creation/update on Sinks, only in sinker update.
