+++ maestro/config_builder_test.go
@@ -1,4 +1,4 @@
-package sinker
+package maestro
 
 import (
 	"context"
