+++ maestro/maestro_service.go
@@ -12,221 +12,28 @@ import (
 	"bufio"
 	"context"
 	"fmt"
-	"os"
-	"os/exec"
-	"strings"
-
 	"github.com/ns1labs/orb/pkg/errors"
 	"go.uber.org/zap"
+	"os"
+	"os/exec"
 )
 
 var (
-	otelCollectorCfg = `---
-receivers:
-  kafka:
-    brokers:
-    - orb-live-stg-kafka.orb-live.svc.cluster.local:9092
-    topic: otlp_metrics-sink-id-222
-    protocol_version: 2.0.0
-extensions:
-  health_check:
-    endpoint: :13133
-    path: /health/status
-    check_collector_pipeline:
-      enabled: true
-      interval: 5m
-      exporter_failure_threshold: 5
-  basicauth/exporter:
-    client_auth:
-      username: admin
-      password: amanda.joaquina
-exporters:
-  prometheusremotewrite:
-    endpoint: https://prometheus.qa.orb.live/api/v1/write
-    auth:
-      authenticator: basicauth/exporter
-service:
-  extensions:
-  - health_check
-  - basicauth/exporter
-  pipelines:
-    metrics:
-      receivers:
-      - kafka
-      exporters:
-      - prometheusremotewrite
-`
-
-	k8sOtelCollector = `
-{
-    "kind": "List",
-    "apiVersion": "v1",
-    "metadata": {},
-    "items": [
-        {
-            "kind": "ConfigMap",
-            "apiVersion": "v1",
-            "metadata": {
-                "name": "otel-collector-config-SINK_ID",
-                "creationTimestamp": null
-            },
-            "data": {
-                "config.yaml": "SINK_CONFIG"
-            }
-        },
-        {
-            "kind": "Deployment",
-            "apiVersion": "apps/v1",
-            "metadata": {
-                "name": "otel-SINK_ID",
-                "creationTimestamp": null,
-                "labels": {
-                    "app": "opentelemetry",
-                    "component": "otel-collector"
-                }
-            },
-            "spec": {
-                "replicas": 1,
-                "selector": {
-                    "matchLabels": {
-                        "app": "opentelemetry",
-                        "component": "otel-collector-SINK_ID"
-                    }
-                },
-                "template": {
-                    "metadata": {
-                        "creationTimestamp": null,
-                        "labels": {
-                            "app": "opentelemetry",
-                            "component": "otel-collector-SINK_ID"
-                        }
-                    },
-                    "spec": {
-                        "volumes": [
-                            {
-                                "name": "varlog",
-                                "hostPath": {
-                                    "path": "/var/log",
-                                    "type": ""
-                                }
-                            },
-                            {
-                                "name": "varlibdockercontainers",
-                                "hostPath": {
-                                    "path": "/var/lib/docker/containers",
-                                    "type": ""
-                                }
-                            },
-                            {
-                                "name": "data",
-                                "configMap": {
-                                    "name": "otel-collector-config-SINK_ID",
-                                    "defaultMode": 420
-                                }
-                            }
-                        ],
-                        "containers": [
-                            {
-                                "name": "otel-collector",
-                                "image": "otel/opentelemetry-collector-contrib:0.60.0",
-                                "resources": {
-                                    "limits": {
-                                        "cpu": "100m",
-                                        "memory": "200Mi"
-                                    },
-                                    "requests": {
-                                        "cpu": "100m",
-                                        "memory": "200Mi"
-                                    }
-                                },
-                                "volumeMounts": [
-                                    {
-                                        "name": "varlog",
-                                        "readOnly": true,
-                                        "mountPath": "/var/log"
-                                    },
-                                    {
-                                        "name": "varlibdockercontainers",
-                                        "readOnly": true,
-                                        "mountPath": "/var/lib/docker/containers"
-                                    },
-                                    {
-                                        "name": "data",
-                                        "readOnly": true,
-                                        "mountPath": "/etc/otelcol-contrib/config.yaml",
-                                        "subPath": "config.yaml"
-                                    }
-                                ],
-                                "terminationMessagePath": "/dev/termination-log",
-                                "terminationMessagePolicy": "File",
-                                "imagePullPolicy": "IfNotPresent"
-                            }
-                        ],
-                        "restartPolicy": "Always",
-                        "terminationGracePeriodSeconds": 30,
-                        "dnsPolicy": "ClusterFirst",
-                        "securityContext": {},
-                        "schedulerName": "default-scheduler"
-                    }
-                },
-                "strategy": {
-                    "type": "RollingUpdate",
-                    "rollingUpdate": {
-                        "maxUnavailable": "25%",
-                        "maxSurge": "25%"
-                    }
-                },
-                "revisionHistoryLimit": 10,
-                "progressDeadlineSeconds": 600
-            },
-            "status": {}
-        },
-        {
-            "kind": "Service",
-            "apiVersion": "v1",
-            "metadata": {
-                "name": "otel-SINK_ID",
-                "creationTimestamp": null,
-                "labels": {
-                    "app": "opentelemetry",
-                    "component": "otel-collector-SINK_ID"
-                }
-            },
-            "spec": {
-                "ports": [
-                    {
-                        "name": "metrics",
-                        "protocol": "TCP",
-                        "port": 8888,
-                        "targetPort": 8888
-                    }
-                ],
-                "selector": {
-                    "component": "otel-collector-SINK_ID"
-                },
-                "type": "ClusterIP",
-                "sessionAffinity": "None"
-            },
-            "status": {
-                "loadBalancer": {}
-            }
-        }
-    ]
-}
-`
 	ErrCreateMaestro   = errors.New("failed to create Otel Collector")
 	ErrConflictMaestro = errors.New("Otel collector already exists")
 )
 
-func (svc maestroService) collectorDeploy(operation string, namespace string, sinkId string, manifest string, config string) error {
-	// prepare manifest
-	manifest = strings.Replace(manifest, "SINK_ID", sinkId, -1)
-	config = strings.Replace(config, "\n", `\n`, -1)
-	manifest = strings.Replace(manifest, "SINK_CONFIG", config, -1)
-	fileContent := []byte(manifest)
-	err := os.WriteFile("/tmp/otel-collector-"+sinkId+".json", fileContent, 0644)
+func (svc maestroService) collectorDeploy(operation, namespace, manifest, sinkId, sinkUrl, sinkUsername, sinkPassword string) error {
+	manifest, err := GetDeploymentJson(sinkId, sinkUrl, sinkUsername, sinkPassword)
 	if err != nil {
-		panic(err)
+		svc.logger.Error("failed to get deployment json", zap.Error(err))
+		return err
+	}
+	fileContent := []byte(manifest)
+	err = os.WriteFile("/tmp/otel-collector-"+sinkId+".json", fileContent, 0644)
+	if err != nil {
+		svc.logger.Error("failed to write file content", zap.Error(err))
+		return err
 	}
 
 	// execute action
@@ -261,8 +68,14 @@ func (svc maestroService) collectorDeploy(operation string, namespace string, si
 	return nil
 }
 
+func (svc maestroService) getConfigFromSinkId(id string) (sinkUrl, sinkUsername, sinkPassword string) {
+
+	return "", "", ""
+}
+
 func (svc maestroService) CreateOtelCollector(ctx context.Context, sinkID string, msg string, ownerID string) error {
-	err := svc.collectorDeploy("apply", "otelcollectors", sinkID, k8sOtelCollector, otelCollectorCfg)
+	sinkUrl, sinkUsername, sinkPassword := svc.getConfigFromSinkId(sinkID)
+	err := svc.collectorDeploy("apply", "otelcollectors", k8sOtelCollector, sinkID, sinkUrl, sinkUsername, sinkPassword)
 	if err != nil {
 		return err
 	}
@@ -270,15 +83,16 @@ func (svc maestroService) CreateOtelCollector(ctx context.Context, sinkID string
 }
 
 func (svc maestroService) UpdateOtelCollector(ctx context.Context, sinkID string, msg string, ownerID string) error {
-	err := svc.collectorDeploy("apply", "otelcollectors", sinkID, k8sOtelCollector, otelCollectorCfg)
+	sinkUrl, sinkUsername, sinkPassword := svc.getConfigFromSinkId(sinkID)
+	err := svc.collectorDeploy("apply", "otelcollectors", k8sOtelCollector, sinkID, sinkUrl, sinkUsername, sinkPassword)
 	if err != nil {
 		return err
 	}
 	return nil
 }
 
-func (svc maestroService) DeleteOtelCollector(ctx context.Context, sinkID string, msg string, ownerID string) error {
-	err := svc.collectorDeploy("delete", "otelcollectors", sinkID, k8sOtelCollector, otelCollectorCfg)
+func (svc maestroService) DeleteOtelCollector(ctx context.Context, sinkID string) error {
+	err := svc.collectorDeploy("delete", "otelcollectors", k8sOtelCollector, sinkID, "", "", "")
 	if err != nil {
 		return err
 	}
